Working code generator to link TUV to DSMACC photolysis IDs
Code reads DSMACC IDs and TUV label from database file and links them
to DSMACC with the help of the TUV input file.
DSMACC FORTRAN code for constants.f90 is generated and saved in a file
in the output folder.
Only the photoreactions currently in the mechanism are outputted.
Checks for TUV flags and broken links.