Further fixes for use of branching ratios
Combine branching ratios in library variables under one index,
redefine array sizes for this.
Introduction of parameter for max. number of branching ratios in params.
Added switch for forced inclusion of inorganic photolysis reactions
in params. Print of inorganic j IDs to tmp file in jmech.
In SR tdlnk, reactions with branching ratios are combinded under the
same index. Call of new SR bratio (in rdfiles.f90).
Adjustment of wrtoutp to combine reactions with branching ratios in the
same CASE.
Fix: Programme now stops, if a j ID is in the mechanism, but not in the
database.
Automatic deletion of tmp file included.
Update of I/O files.
Update of README.