Allow printing of all photolysis reactions from the database, fixes
Added option to print all possible photolysis reactions in output.
Make LOGICAL inorg to INTEGER mech, with 3 options for different subsets
of mechanism output. Adjustments in print statements to the temporary
file in jmech.
FIX: Adjustments/Fixes in database and TUV input files for new reactions
     of HPALD and reactions with branching ratio of 1 (see README.md).
FIX: Output excludes factor of 1.0 in reactions with this branching
     ratio.
Update of README.md.