Making folder paths optional, updating structure
Creating a database (DB) folder and moving database and TUV files to it.
Updating I/O files.
In subs/finit.f90, adjusting folder paths for new DB folder and making
folder paths an optional argument. Files have to be stored in the 
assigned folder, but it paths are obsolete in the programme arguments,
default paths will be used.