Adjust array sizes in constants.f90
Introduction of new variable fconst for file name and folder path of 
the DSMACC routine constants.f90 in linkphot.f90 and finit.f90.
In finit, retrieving file name and path for constants file from 5th
programme argument, if available. Updated screen output.
In tdlnk in rdfiles.f90, including empty lines in cycle command when
reading the database file. This will allow the programme to skip over 
empty lines and comment lines as well.
In wrtoutp in wrtoutp.f90, introduction of parameter jmax, which stores
the maximum value of DSMACC j labels. Initialisation of jmax at the
beginning, determination in main loop and rounding to the next full 100
at the end of the routine. Return of jmax to linkphot.
New routine adjARRsiz in wrtoutp.f90 that calls perl to replace the 
maximum array size of all j related arrays in the DSMACC routine 
constants.f90, if the file is specified in the 5. programme argument.
Updating line length parameter llen to 120 in MOD/params.f90.
Updating I/O and database files.
Updated README and gitignore.